requirements.txt
Pillow
python-pptx 